Adjust the configurations
The Debug now runs santiziers and coverage. The RelWithDebInfo now
runs valgrind.

The reason for this change is a preparation for perf. We do not want
to instrument santiziers and coverage into the executable when
profiling with perf. The reason for not using the debug configuration
is that we want perf to run on something that is as close as release
as possible but with debug symbols.