Fix includes
Previously all directories under src were available for all targets
as there was an include_directories in the top CMakeLists file. This
made the “include scope" unnecessary big as some targets got an
include path that contained more than they actually needed. By
removing include_directories in the top CMakeLists file we make sure
that a target can only include things that it is actually needed.

Another problem was that all libraries did
    target_include_directories(
        LinearAlgebra PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
which resulted in that all header files were added to the include
path, even private ones. Another drawback was that a user of the
library could include a header file without any prefix like
“header.h” instead of “Library/header.h”. This can cause name
conflicts and the source of headers is much clearer if we enforce
the include to be “Library/header.h”. To accomplish this we move
the public header files to include/Library and add the include
directory as the directory for target_include_directories.
With this approach a user of a library must use the library prefix
when including header files and only public header files will be
available.

All “”-includes that are not relative includes are also changed to
<>-includes  for clarity.